+++ internal/adapter/sqlite/note_index.go
@@ -17,6 +17,7 @@ type NoteIndex struct {
 
 type dao struct {
 	notes       *NoteDAO
+	links       *LinkDAO
 	collections *CollectionDAO
 	metadata    *MetadataDAO
 }
@@ -41,6 +42,15 @@ func (ni *NoteIndex) Find(opts core.NoteFindOpts) (notes []core.ContextualNote, 
 func (ni *NoteIndex) FindMinimal(opts core.NoteFindOpts) (notes []core.MinimalNote, err error) {
 	err = ni.commit(func(dao *dao) error {
 		notes, err = dao.notes.FindMinimal(opts)
+		return err
+	})
+	return
+}
+
+// FindLinksBetweenNotes implements core.NoteIndex.
+func (ni *NoteIndex) FindLinksBetweenNotes(ids []core.NoteID) (links []core.ResolvedLink, err error) {
+	err = ni.commit(func(dao *dao) error {
+		links, err = dao.links.FindBetweenNotes(ids)
 		return err
 	})
 	return
@@ -73,6 +83,11 @@ func (ni *NoteIndex) Add(note core.Note) (id core.NoteID, err error) {
 			return err
 		}
 
+		err = ni.addLinks(dao, id, note)
+		if err != nil {
+			return err
+		}
+
 		return ni.associateTags(dao.collections, id, note.Tags)
 	})
 
@@ -83,17 +98,27 @@ func (ni *NoteIndex) Add(note core.Note) (id core.NoteID, err error) {
 // Update implements core.NoteIndex.
 func (ni *NoteIndex) Update(note core.Note) error {
 	err := ni.commit(func(dao *dao) error {
-		noteId, err := dao.notes.Update(note)
+		id, err := dao.notes.Update(note)
 		if err != nil {
 			return err
 		}
 
-		err = dao.collections.RemoveAssociations(noteId)
+		// Reset links
+		err = dao.links.RemoveAll(id)
+		if err != nil {
+			return err
+		}
+		err = ni.addLinks(dao, id, note)
 		if err != nil {
 			return err
 		}
 
-		return ni.associateTags(dao.collections, noteId, note.Tags)
+		// Reset tags
+		err = dao.collections.RemoveAssociations(id)
+		if err != nil {
+			return err
+		}
+		return ni.associateTags(dao.collections, id, note.Tags)
 	})
 
 	return errors.Wrapf(err, "%v: failed to update note index", note.Path)
@@ -112,6 +137,40 @@ func (ni *NoteIndex) associateTags(collections *CollectionDAO, noteId core.NoteI
 	}
 
 	return nil
+}
+
+func (ni *NoteIndex) addLinks(dao *dao, id core.NoteID, note core.Note) error {
+	links, err := ni.resolveLinkNoteIDs(dao, id, note.Links)
+	if err != nil {
+		return err
+	}
+
+	err = dao.links.Add(links)
+	if err != nil {
+		return err
+	}
+
+	return dao.links.SetTargetID(note.Path, id)
+}
+
+func (ni *NoteIndex) resolveLinkNoteIDs(dao *dao, sourceID core.NoteID, links []core.Link) ([]core.ResolvedLink, error) {
+	resolvedLinks := []core.ResolvedLink{}
+
+	for _, link := range links {
+		allowPartialMatch := (link.Type == core.LinkTypeWikiLink)
+		targetID, err := dao.notes.FindIdByHref(link.Href, allowPartialMatch)
+		if err != nil {
+			return resolvedLinks, err
+		}
+
+		resolvedLinks = append(resolvedLinks, core.ResolvedLink{
+			Link:     link,
+			SourceID: sourceID,
+			TargetID: targetID,
+		})
+	}
+
+	return resolvedLinks, nil
 }
 
 // Remove implements core.NoteIndex
@@ -162,6 +221,7 @@ func (ni *NoteIndex) commit(transaction func(dao *dao) error) error {
 		return ni.db.WithTransaction(func(tx Transaction) error {
 			dao := dao{
 				notes:       NewNoteDAO(tx, ni.logger),
+				links:       NewLinkDAO(tx, ni.logger),
 				collections: NewCollectionDAO(tx, ni.logger),
 				metadata:    NewMetadataDAO(tx),
 			}