+++ internal/core/link.go
@@ -3,21 +3,30 @@ package core
 // Link represents a link in a note to another note or an external resource.
 type Link struct {
 	// Label of the link.
-	Title string
+	Title string `json:"title"`
 	// Destination URI of the link.
-	Href string
+	Href string `json:"href"`
 	// Type of link, e.g. wiki link.
-	Type LinkType
+	Type LinkType `json:"type"`
 	// Indicates whether the target is a remote (e.g. HTTP) resource.
-	IsExternal bool
+	IsExternal bool `json:"isExternal"`
 	// Relationships between the note and the linked target.
-	Rels []LinkRelation
+	Rels []LinkRelation `json:"rels"`
 	// Excerpt of the paragraph containing the note.
-	Snippet string
+	Snippet string `json:"snippet"`
 	// Start byte offset of the snippet in the note content.
-	SnippetStart int
+	SnippetStart int `json:"snippetStart"`
 	// End byte offset of the snippet in the note content.
-	SnippetEnd int
+	SnippetEnd int `json:"snippetEnd"`
+}
+
+// ResolvedLink represents a link between two indexed notes.
+type ResolvedLink struct {
+	Link
+	SourceID   NoteID `json:"sourceId"`
+	SourcePath string `json:"sourcePath"`
+	TargetID   NoteID `json:"targetId"`
+	TargetPath string `json:"targetPath"`
 }
 
 // LinkType represents the kind of link, e.g. wiki link.