+++ internal/core/notebook.go
@@ -236,6 +236,11 @@ func (n *Notebook) FindMatching(terms string) (*MinimalNote, error) {
 	})
 }
 
+// FindLinksBetweenNotes retrieves the links between the given notes.
+func (n *Notebook) FindLinksBetweenNotes(ids []NoteID) ([]ResolvedLink, error) {
+	return n.index.FindLinksBetweenNotes(ids)
+}
+
 // FindCollections retrieves all the collections of the given kind.
 func (n *Notebook) FindCollections(kind CollectionKind, sorters []CollectionSorter) ([]Collection, error) {
 	return n.index.FindCollections(kind, sorters)